requirements.txt
telethon==1.36.0
pyyaml==6.0.2
beautifulsoup4==4.12.3
python-dotenv==1.0.1